requirements.txt
fastapi
uvicorn
pywebview
pyaudiowpatch
websockets
groq
anthropic
python-dotenv
